# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminOrderPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminOrderSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for older orders; null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AdminOrderPage",
        "type": "object"
      },
...
      "AdminProductPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminProductRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AdminProductPage",
        "type": "object"
      },
...
      "AdminStatusChange": {
        "additionalProperties": false,
        "description": "Fulfilment steps. Payment, cancellation and refunds have their own paths.",
        "properties": {
          "to": {
            "enum": [
              "fulfilling",
              "shipped",
              "delivered"
            ],
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "to"
        ],
        "title": "AdminStatusChange",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The envelope every error response uses.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InventoryResponse": {
        "properties": {
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity_available": {
            "title": "Quantity Available",
            "type": "integer"
          },
          "quantity_on_hand": {
            "title": "Quantity On Hand",
            "type": "integer"
          },
          "quantity_reserved": {
            "title": "Quantity Reserved",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "description": "Also returned as the ETag header. Send it back in If-Match to update.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "product_id",
          "quantity_on_hand",
          "quantity_reserved",
          "quantity_available",
          "version",
          "updated_at"
        ],
        "title": "InventoryResponse",
        "type": "object"
      },
...
      "OrderStatus": {
        "enum": [
          "pending_payment",
          "paid",
          "fulfilling",
          "shipped",
          "delivered",
          "cancelled",
          "refunded"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
...
      "ProductUpdate": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Is Active"
          },
          "price_cents": {
            "anyOf": [
              {
                "maximum": 2147483647,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Price in minor units (integer, never a float).",
            "title": "Price Cents"
          }
        },
        "title": "ProductUpdate",
        "type": "object"
      },
...
      "RefundRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reason"
          }
        },
        "title": "RefundRequest",
        "type": "object"
      },
...
      "StockUpdate": {
        "additionalProperties": false,
        "description": "Only stock on hand is editable. Reserved stock belongs to checkout and changes only through it.",
        "properties": {
          "quantity_on_hand": {
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Quantity On Hand",
            "type": "integer"
          }
        },
        "required": [
          "quantity_on_hand"
        ],
        "title": "StockUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "PC hardware marketplace with a database-enforced build compatibility engine.",
    "title": "Forge API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/inventory/{product_id}": {
      "get": {
        "description": "",
        "operationId": "get__api_v1_admin_inventory_{product_id}",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Current stock levels. The ETag header carries the version required to update them.",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "patch__api_v1_admin_inventory_{product_id}",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Precondition Failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Set stock on hand. Requires If-Match with the current ETag (optimistic concurrency): 412 if the stock changed since it was read, 428 if If-Match is missing, 409 if the new level would fall below stock already reserved by checkouts.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "description": "",
        "operationId": "get__api_v1_admin_orders",
        "parameters": [
          {
            "description": "",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Cursor"
            }
          },
          {
            "description": "",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderPage"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "All orders, newest first, optionally filtered by status, with each order's customer and the actions an administrator may take next.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_number}": {
      "get": {
        "description": "",
        "operationId": "get__api_v1_admin_orders_{order_number}",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "order_number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetail"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Any order, with the customer's email.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_number}/refund": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_admin_orders_{order_number}_refund",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "order_number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetail"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Refund the order's payment in full through Stripe. 409 if the order cannot be refunded in its current state (checked before any money moves); 502 if Stripe did not complete it.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_number}/status": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_admin_orders_{order_number}_status",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "order_number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDetail"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Move a paid order through fulfilment: fulfilling, shipped, delivered. The database rejects illegal jumps (409 invalid_status_transition). Recorded in the status history with the admin.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products": {
      "get": {
        "description": "",
        "operationId": "get__api_v1_admin_products",
        "parameters": [
          {
            "description": "",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Kind"
            }
          },
          {
            "description": "Name or SKU contains.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Name or SKU contains.",
              "title": "Q"
            }
          },
          {
            "description": "",
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Active"
            }
          },
          {
            "description": "",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductPage"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "All products, including inactive ones, with stock levels. Keyset-paged by id.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{product_id}": {
      "patch": {
        "description": "",
        "operationId": "patch__api_v1_admin_products_{product_id}",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change a product's price or availability. Price changes are recorded in price_history by a trigger.",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "security": [],
...
}